refactor(#281) : 리퀘스트 작성
