+++ src/main/java/odyssey/backend/presentation/ai/dto/request/vo/CourseRequest.java
@@ -1,16 +1,18 @@
-package odyssey.backend.presentation.ai.dto.request;
+package odyssey.backend.presentation.ai.dto.request.vo;
 
 import jakarta.validation.constraints.NotBlank;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
+@Getter
 @NoArgsConstructor
 @AllArgsConstructor
-@Getter
-public class GenerateRequest {
+public class CourseRequest {
 
     @NotBlank(message = "필수값입니다.")
-    private String prompt;
+    private String title;
 
+    @NotBlank(message = "필수값입니다.")
+    private String description;
 }
